refactor: styled component 위치 하단으로 이동
