+++ src/components/FieldTitle.tsx
@@ -1,17 +1,5 @@
 import styled from 'styled-components';
 
-const Title = styled.h1`
-  color: #000;
-  font-size: 18px;
-  margin-bottom: 4px;
-`;
-
-const SubTitle = styled.p`
-  color: #8b95a1;
-  font-size: 9.5px;
-  margin-bottom: 16px;
-`;
-
 interface Props {
   title: string;
   subtitle?: string;
@@ -25,3 +13,15 @@ export default function FieldTitle({ title, subtitle }: Props) {
     </>
   );
 }
+
+const Title = styled.h1`
+  color: #000;
+  font-size: 18px;
+  margin-bottom: 4px;
+`;
+
+const SubTitle = styled.p`
+  color: #8b95a1;
+  font-size: 9.5px;
+  margin-bottom: 16px;
+`;
